Use client less connection when scanning MultiBranch.
