+++ src/main/java/org/jenkinsci/plugins/p4/client/ClientHelper.java
@@ -25,7 +25,6 @@ import com.perforce.p4java.option.client.RevertFilesOptions;
 import com.perforce.p4java.option.client.SyncOptions;
 import com.perforce.p4java.option.server.ChangelistOptions;
 import com.perforce.p4java.option.server.GetChangelistsOptions;
-import com.perforce.p4java.option.server.GetDepotFilesOptions;
 import com.perforce.p4java.option.server.GetFileContentsOptions;
 import com.perforce.p4java.option.server.OpenedFilesOptions;
 import com.perforce.p4java.server.CmdSpec;
@@ -995,26 +994,6 @@ public class ClientHelper extends ConnectionHelper {
 	}
 
 	/**
-	 * Get the latest change on the given path
-	 *
-	 * @param path Perforce depot path //foo/...
-	 * @return change number
-	 * @throws Exception push up stack
-	 */
-	public long getHead(String path) throws Exception {
-		List<IFileSpec> spec = FileSpecBuilder.makeFileSpecList(path);
-
-		GetChangelistsOptions opts = new GetChangelistsOptions();
-		opts.setMaxMostRecent(1);
-
-		List<IChangelistSummary> changes = connection.getChangelists(spec, opts);
-		if (!changes.isEmpty()) {
-			return changes.get(0).getId();
-		}
-		return -1;
-	}
-
-	/**
 	 * Gets the Changelist (p4 describe -s); shouldn't need a client, but
 	 * p4-java throws an exception if one is not set.
 	 *
@@ -1296,12 +1275,5 @@ public class ClientHelper extends ConnectionHelper {
 
 	public IClient getClient() {
 		return iclient;
-	}
-
-	public boolean hasFile(String depotPath) throws Exception {
-		List<IFileSpec> files = FileSpecBuilder.makeFileSpecList(depotPath);
-		GetDepotFilesOptions opts = new GetDepotFilesOptions();
-		List<IFileSpec> specs = connection.getDepotFiles(files, opts);
-		return validate.checkCatch(specs, "");
 	}
 }
